pyproject.toml
[project]
name = "web13hm"
version = "0.1.0"
description = ""
authors = [
    {name = "Taras",email = "93609222+aps06@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "fastapi-jwt-auth (>=0.5.0,<0.6.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (==1.7.4)",
    "bcrypt (<4.0)",
    "fastapi-mail (>=1.5.7,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "web13hm", from = "src"}]


[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
